release: v1.1.0

+++ backend/src/main/java/com/yagubogu/stat/dto/OpponentWinRateRow.java
@@ -5,8 +5,8 @@ public record OpponentWinRateRow(
         String name,
         String shortName,
         String teamCode,
-        long wins,
-        long losses,
-        long draws
+        int wins,
+        int losses,
+        int draws
 ) {
 }
